+++ include/rviz_auto_sensor_panel/types.hpp
@@ -12,7 +12,9 @@ enum class SensorCategory
   Unknown,
   Lidar,
   Camera,
-  PointCloud
+  PointCloud,
+  Navigation,
+  Visualization
 };
 
 struct DiscoveredTopic